@@ -3,12 +3,12 @@
   "version": "1.0.0",
   "description": "",
   "engines": {
-    "node": "14.x"
+    "node": "16.x"
   },
+  "type": "module",
   "main": "app.js",
   "scripts": {
-    "start": "node app.js",
-    "dev": "nodemon app.js --config config.json",
+    "start": "node app.js --config config.json",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
