@@ -1,14 +1,12 @@
-const { MongoClient } = require('mongodb')
+import { MongoClient } from 'mongodb'
 
 const password = process.env.MONGO_PASSWORD
 
 const uri = `mongodb+srv://admin:${password}@cluster0.jwoag.mongodb.net/sorcsapp?retryWrites=true&w=majority`
 
-function getClient() {
+export default function getClient() {
     return new MongoClient(uri, {
         useNewUrlParser: true,
         useUnifiedTopology: true,
     })
 }
-
-module.exports = getClient
